Main script for creating, and viewing playlists
This script's function is to search through the api, and retrieve individual songs back from it. After which, it stores the query in a playlist specified by the user. This is done by the writing the playlist name, followed by a colon, then the name of the song. 
(Ex. playlistname:songname). Duplicate entries are dismissed.