Tempory file for apiScipt operations
This script is used to save the number of pages returned from the api after a query. If the number of pages is less than 150, the query is most likely accurate and returns a singular artist/album. If the number of pages is more than 150, its safe to assume the query wasn't accurate enough and will return the names of all similar queries. This operation is done in apiScript.php